% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as_tibble.metadset}
\alias{as_tibble.metadset}
\title{Coerce a \code{metadset} to tibble}
\usage{
\method{as_tibble}{metadset}(x, ...)
}
\arguments{
\item{x}{A \code{metadset}.}

\item{...}{Ignored.}
}
\value{
A tibble with at least \code{study_id}, \code{yi}, \code{vi}.
}
\description{
Returns the standardized study-level data.
}
