% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a dataset from common meta-analysis packages}
\usage{
get_dataset(name, package = c("metadat", "metafor", "meta", "dmetar"))
}
\arguments{
\item{name}{Dataset object name (e.g., "dat.bcg").}

\item{package}{One of c("metadat","metafor","meta","dmetar").}
}
\value{
The dataset object.
}
\description{
Get a dataset from common meta-analysis packages
}
