% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{list_datasets}
\alias{list_datasets}
\title{List available datasets from the registry}
\usage{
list_datasets(pattern = NULL)
}
\arguments{
\item{pattern}{Optional regex to filter by id or title}
}
\description{
List available datasets from the registry
}
