% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{meta_forest}
\alias{meta_forest}
\title{Forest plot for a meta-analysis fit}
\usage{
meta_forest(fit, ...)
}
\arguments{
\item{fit}{A \code{metafit} object or a \code{metafor::rma.uni} object.}

\item{...}{Passed to \code{\link[metafor:forest]{metafor::forest()}}.}
}
\description{
Forest plot for a meta-analysis fit
}
