% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{meta_funnel}
\alias{meta_funnel}
\title{Funnel plot for a meta-analysis fit}
\usage{
meta_funnel(fit, ...)
}
\arguments{
\item{fit}{A \code{metafit} object or a \code{metafor::rma.uni} object.}

\item{...}{Passed to \code{\link[metafor:funnel]{metafor::funnel()}}.}
}
\description{
Funnel plot for a meta-analysis fit
}
