% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{meta_loo}
\alias{meta_loo}
\title{Leave-one-out diagnostics for a meta-analysis fit}
\usage{
meta_loo(fit, ...)
}
\arguments{
\item{fit}{A \code{metafit} object or a \code{metafor::rma.uni} object.}

\item{...}{Passed to \code{\link[metafor:leave1out]{metafor::leave1out()}}.}
}
\description{
Leave-one-out diagnostics for a meta-analysis fit
}
