% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_metadset}
\alias{new_metadset}
\title{Create a standardized meta-analysis dataset object}
\usage{
new_metadset(data, meta = list(), id = NULL, schema_version = "1.0.0")
}
\arguments{
\item{data}{A data.frame/tibble with columns \code{study_id}, \code{yi}, \code{vi}.}

\item{meta}{A named list with metadata (source, citation, license, tags, etc.)}

\item{id}{Optional registry id}

\item{schema_version}{Schema version}
}
\value{
An object of class \code{metadset}
}
\description{
Create a standardized meta-analysis dataset object
}
