% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_columns}
\alias{standardize_columns}
\title{Standardize raw dataset columns into (study_id, yi, vi)}
\usage{
standardize_columns(df, map)
}
\arguments{
\item{df}{Raw data.frame}

\item{map}{Named list mapping raw columns to standard names.
Recognized: study_id, yi, vi, sei, escalc (list of args for metafor::escalc)}
}
\value{
Tibble with standardized columns
}
\description{
Standardize raw dataset columns into (study_id, yi, vi)
}
