% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_metadset}
\alias{validate_metadset}
\title{Validate a metadset}
\usage{
validate_metadset(x)
}
\arguments{
\item{x}{A \code{metadset}}
}
\value{
The input (invisibly) if valid; otherwise errors.
}
\description{
Validate a metadset
}
