% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets known to metahub}
\usage{
list_datasets()
}
\value{
data.frame with columns: id, package, object, title
}
\description{
List datasets known to metahub
}
